MIN-41-close unused heredoc pipes + starting signals

+++ SRCS/free.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tgellon <tgellon@student.42lyon.fr>        +#+  +:+       +#+        */
+/*   By: rrebois <rrebois@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/05 12:33:48 by rrebois           #+#    #+#             */
-/*   Updated: 2023/05/16 10:52:41 by tgellon          ###   ########lyon.fr   */
+/*   Updated: 2023/05/17 09:32:04 by rrebois          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ void	free_all(t_data *data)
 
 void	free_hd_struct(t_data *data)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	ft_free_pp(data->hd->LIMITER);
